# server/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend API for the SITCON Camp 2026 game.",
        "title": "Camp 2026 Game API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/api",
    "paths": {
        "/admin/login": {
            "post": {
                "description": "Uses the ADMIN_PASSWORD environment variable to create an admin session cookie.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Login as admin",
                "parameters": [
                    {
                        "description": "Admin login request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/admin.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    }
                }
            }
        },
...
        "/admin/settings": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get admin game settings",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/admin.SettingsResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Update admin game settings",
                "parameters": [
                    {
                        "description": "Admin settings request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.SettingsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/admin.SettingsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "User-facing endpoint. Validates the issued game URL token, rotates it, and writes the fresh session token to the camp2026_auth cookie.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Authentication"
                ],
                "summary": "User login with auth token",
                "parameters": [
                    {
                        "description": "Auth login request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apimodel.AuthLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apimodel.AuthLoginResponse"
                        },
                        "headers": {
                            "Set-Cookie": {
                                "type": "string",
                                "description": "camp2026_auth=\u003cauth-token\u003e; Path=/; HttpOnly; Secure; SameSite=Strict"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "security": [
                    {
                        "AuthCookieAuth": []
                    }
                ],
                "description": "Clears the camp2026_auth cookie for the current browser session.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Authentication"
                ],
                "summary": "User logout",
                "responses": {
                    "204": {
                        "description": "No Content",
                        "headers": {
                            "Set-Cookie": {
                                "type": "string",
                                "description": "camp2026_auth=; Path=/; HttpOnly; Secure; SameSite=Strict; Max-Age=0"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "admin.LoginRequest": {
            "type": "object",
            "required": [
                "password"
            ],
            "properties": {
                "password": {
                    "type": "string"
                }
            }
        },
        "admin.LoginResponse": {
            "type": "object",
            "properties": {
                "authenticated": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "admin.SettingsRequest": {
            "type": "object",
            "required": [
                "computerBattlesEnabled",
                "computerEasyAccuracy",
                "computerHardAccuracy",
                "computerNormalAccuracy"
            ],
            "properties": {
                "computerBattlesEnabled": {
                    "type": "boolean"
                },
                "computerEasyAccuracy": {
                    "type": "integer",
                    "maximum": 100,
                    "minimum": 0
                },
                "computerHardAccuracy": {
                    "type": "integer",
                    "maximum": 100,
                    "minimum": 0
                },
                "computerNormalAccuracy": {
                    "type": "integer",
                    "maximum": 100,
                    "minimum": 0
                }
            }
        },
        "admin.SettingsResponse": {
            "type": "object",
            "properties": {
                "computerBattlesEnabled": {
                    "type": "boolean"
                },
                "computerEasyAccuracy": {
                    "type": "integer"
                },
                "computerHardAccuracy": {
                    "type": "integer"
                },
                "computerNormalAccuracy": {
                    "type": "integer"
                }
            }
        },
        "apimodel.AuthLoginRequest": {
            "type": "object",
            "required": [
                "token"
            ],
            "properties": {
                "token": {
                    "type": "string",
                    "maxLength": 512,
                    "example": "token-1"
                }
            }
        },
        "apimodel.AuthLoginResponse": {
            "type": "object",
            "properties": {
                "player": {
                    "$ref": "#/definitions/apimodel.AuthPlayerSummary"
                }
            }
        },
        "apimodel.AuthPlayerSummary": {
            "type": "object",
            "properties": {
                "avatarUrl": {
                    "type": "string",
                    "example": "https://example.test/avatar/alice.png"
                },
                "nickname": {
                    "type": "string",
                    "example": "Alice"
                },
                "openPower": {
                    "type": "integer",
                    "example": 1280
                },
                "playerId": {
                    "type": "string",
                    "example": "7H9K2Q"
                },
                "role": {
                    "type": "string",
                    "example": "staff"
                },
                "team": {
                    "$ref": "#/definitions/apimodel.AuthTeamSummary"
                }
            }
        },
        "apimodel.AuthTeamSummary": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Blue Team"
                },
                "teamId": {
                    "type": "string",
                    "example": "8M4RXP"
                }
            }
        },
...
        "httpx.ErrorDetail": {
            "type": "object",
            "properties": {
                "location": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "httpx.ProblemDetails": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "detail": {
                    "type": "string"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/httpx.ErrorDetail"
                    }
                },
                "instance": {
                    "type": "string"
                },
                "requestId": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AuthCookieAuth": {
            "description": "User auth cookie. Send as `Cookie: camp2026_auth=\u003cbase64url-auth-token\u003e`.",
            "type": "apiKey",
            "name": "Cookie",
            "in": "header"
        }
    }
}